:bug: cycle 린트 오류를 해결한다
