+++ components/Browser/index.ts
@@ -1,1 +1,2 @@
 export { default as Browser } from './Main';
+export type { ProjectInterface } from './types';
